Update with newest Instant Client versions
Upgrade 19.8 to 19.9, and add new version 21.1

Remove version 11 and 12 image generation, as I'm unable to download the drivers from Oracle due to some "technical difficulty" on their site.
Old images are still available on Docker Hub, but I will not be generating new ones -- at least for now.

The new v21 drivers have a slightly different file naming scheme, so I chose to rename the older drivers to match and update the template.